requirements.txt
PyMuPDF==1.23.8
python-dotenv==1.0.0
numpy==1.24.3
faiss-cpu==1.7.4
openai==1.3.0
pytesseract==0.3.10
Pillow==10.0.0
python-docx==0.8.11
beautifulsoup4==4.12.2
pandas==2.0.3
lxml==4.9.3
python-pptx==0.6.23
openpyxl==3.1.2
pdfplumber==0.10.3
